requirements.txt
odoo_upgrade @ git+https://github.com/odoo/upgrade-util@master
qrcode